Add files via upload
This is the new Shell with Exit Command. If we have to demo, change the directory of the History file which is just a basic text file on the shell2 file (currently has my Linux directory link). The History file currently contains all the tests I conducted. We can clear it if needed.